refactor ink component to compatible with taskResults input
