+++ packages/checkup-formatter-pretty/__tests__/bar-test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from 'ink-testing-library';
-import { Bar } from '../src/components/bar';
+import { Bar } from '../src/components/utils/bar';
 import { BarData } from '../src/types';
 
 const stripAnsi = require('strip-ansi');
