+++ packages/checkup-formatter-pretty/__tests__/list-test.tsx
@@ -1,21 +1,22 @@
 import * as React from 'react';
 import { render } from 'ink-testing-library';
-import { ListItem } from '../src/components/list-item';
-import { List } from '../src/components/list';
+import { Text } from 'ink';
+import { List } from '../src/components/utils/list';
 
 describe('Test list component', () => {
   it('can generate list component', async () => {
+    const data = ['item1', 'item2', 'item3'];
     const { stdout } = render(
       <List>
-        <ListItem data="list 1"></ListItem>
-        <ListItem data="list 2"></ListItem>
-        <ListItem data="list 3"></ListItem>
+        {data.map((item) => {
+          return <Text key={item}>{item}</Text>;
+        })}
       </List>
     );
     expect(stdout.lastFrame()).toMatchInlineSnapshot(`
-      "list 1
-      list 2
-      list 3"
+      "item1
+      item2
+      item3"
     `);
   });
 });
