+++ packages/checkup-formatter-pretty/src/component-provider.ts
@@ -6,9 +6,11 @@ export const registeredComponents = new Map<string, React.FC>();
 
 export function registerDefaultComponents(): Map<string, React.FC> {
   let builtInComponents = new Set(
-    readdirSync(join(__dirname, 'components')).map((file) => {
-      return parse(file).base.split('.')[0];
-    })
+    readdirSync(join(__dirname, 'components'), { withFileTypes: true })
+      .filter((file) => file.isFile())
+      .map((file) => {
+        return parse(file.name).base.split('.')[0];
+      })
   );
 
   for (let component of builtInComponents) {
