--- packages/checkup-formatter-pretty/src/components/list-item.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import { Text, Box } from 'ink';
-
-export const ListItem: React.FC<{ data: any }> = ({ data }) => {
-  return (
-    <Box>
-      <Text>{data}</Text>
-    </Box>
-  );
-};
