+++ packages/checkup-formatter-pretty/src/components/task-result-list.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { Text } from 'ink';
+import { RuleResults } from '@checkup/core';
+
+export const TaskResultList: React.FC<{ taskResult: RuleResults }> = ({ taskResult }) => {
+  return (
+    <Text>
+      ruleId: {taskResult.rule.id} result(value): {taskResult.results.length}
+    </Text>
+  );
+};
