+++ packages/checkup-formatter-pretty/src/components/utils/list.tsx
